Relocate azure_pipelines starting point and rename CI references
